Move potions out of other
also adds mana potions
